consolidated into one build
Instead of two separate builds for React and Preact, this now has only
one build that supports both. This simplifies usage when aliasing React
with Preact through Webpack and the like.

Added Emerge as peer dependency.

The source file is now `.mjs` and can be imported directly in newer Node
versions.